Move `build` to `lib` folder
